Make more helper functions.
