+++ Cargo.toml
@@ -15,12 +15,17 @@ exclude = [
 ]
 
 [dependencies]
-serde = { version = "1.0", features = ["derive", "serde_derive"] }
-serde_json = { version = "1.0" }
-serde_path_to_error = "0.1.17"
+serde = { version = "1.0", features = ["derive", "serde_derive"], optional = true }
+serde_json = { version = "1.0", optional = true }
+serde_path_to_error = { version = "0.1.17", optional = true }
+chrono = { version = "0.4", features = ["serde"], optional = true }
 thiserror = "2.0.12"
-chrono = { version = "0.4", features = ["serde"] }
 
 [dev-dependencies]
 image = "0.25.6"
 walkdir = "2.5.0"
+
+[features]
+default = ["full-deserialize"]
+full-deserialize = ["tag-deserialize", "serde_path_to_error", "chrono"]
+tag-deserialize = ["serde", "serde_json"]
