+++ src/core/mod.rs
@@ -1,0 +1,6 @@
+mod base_helpers;
+pub mod exiftool;
+mod helpers;
+
+#[cfg(feature = "tag-deserialize")]
+mod serde_helpers;
